creating methods for adding and removing users
