--- concert_compainion_flutter/lib/Swipe.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SwipeScreen extends StatelessWidget {
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Concert Companion"),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.people),
-            onPressed: () {
-              Navigator.pushNamed(context, '/swipe');
-            },
-          ),
-          // action button
-          IconButton(
-            icon: Icon(Icons.message),
-            onPressed: () {
-              Navigator.pushNamed(context, '/message');
-            },
-          ),
-        ],
-      ),
-      body: Stack(
-
-      ),
-    );
-  }
-
-}
