--- concert_compainion_flutter/lib/profile_card_alignment.dart
@@ -1,51 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileCardAlignment extends StatelessWidget {
-  final int cardNum;
-  ProfileCardAlignment(this.cardNum);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: Stack(
-        children: <Widget>[
-          SizedBox.expand(
-            child: Material(
-              borderRadius: BorderRadius.circular(12.0),
-              child: Image.asset('res/portrait.jpeg', fit: BoxFit.cover),
-            ),
-          ),
-          SizedBox.expand(
-            child: Container(
-              decoration: BoxDecoration(
-                  gradient: LinearGradient(
-                      colors: [Colors.transparent, Colors.black54],
-                      begin: Alignment.center,
-                      end: Alignment.bottomCenter)),
-            ),
-          ),
-          Align(
-            alignment: Alignment.bottomLeft,
-            child: Container(
-                padding: EdgeInsets.symmetric(vertical: 16.0, horizontal: 16.0),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Text('Card number $cardNum',
-                        style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 20.0,
-                            fontWeight: FontWeight.w700)),
-                    Padding(padding: EdgeInsets.only(bottom: 8.0)),
-                    Text('A short description.',
-                        textAlign: TextAlign.start,
-                        style: TextStyle(color: Colors.white)),
-                  ],
-                )),
-          )
-        ],
-      ),
-    );
-  }
-}
