+++ concert_compainion_flutter/lib/swipe_feed_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'cards_section_alignment.dart';
 import 'cards_section_draggable.dart';
+import 'profile_card_draggable.dart';
 
 class SwipeFeedPage extends StatefulWidget {
   @override
@@ -8,7 +8,28 @@ class SwipeFeedPage extends StatefulWidget {
 }
 
 class _SwipeFeedPageState extends State<SwipeFeedPage> {
-  bool showAlignmentCards = false;
+
+  List<ProfileCardDraggable> cards = List();
+  int cardsCounter = 0;
+
+  void changeCardsOrder() {
+    setState(() {
+      cards[0] = cards[1];
+      cards[1] = cards[2];
+
+      cards[2] = ProfileCardDraggable(cardsCounter);
+      cardsCounter++;
+    });
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    for (cardsCounter = 0; cardsCounter < 3; cardsCounter++) {
+      cards.add(ProfileCardDraggable(cardsCounter));
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -33,9 +54,7 @@ class _SwipeFeedPageState extends State<SwipeFeedPage> {
       ),
       body: Column(
         children: <Widget>[
-          showAlignmentCards
-              ? CardsSectionAlignment(context)
-              : CardsSectionDraggable(),
+          CardsSectionDraggable(),
           buttonsRow()
         ],
       ),
@@ -64,7 +83,6 @@ class _SwipeFeedPageState extends State<SwipeFeedPage> {
               heroTag: "btn3",
               onPressed: () {
                 //api post call to report wanted friend
-                //also update ui to remove current card
               },
               backgroundColor: Colors.white,
               child: Icon(Icons.favorite, color: Colors.green, size: 40,),
